* arts-1-14-184
* src/constants.cc:

Remove Joker definition here, it belongs in the matpack library.

Include mystring.h instead of arts.h.

* src/matpackI.cc: Add definition and initialization for global joker.

* src/matpackI.h: Make global joker object constant.

* src/agenda_class.cc (Agenda::append): Remove _U_ from ws.

* src/libmicrohttpd/connection.c, src/libmicrohttpd/daemon.c:

Add cast to unsigned int to silence some warnings.


git-svn-id: https://www.sat.ltu.se/svn/rt/arts/trunk@6878 aaf1aab0-4228-0410-ad68-8dceda47f409

+++ src/libmicrohttpd/connection.c
@@ -1385,7 +1385,7 @@ process_request_body (struct MHD_Connection *connection)
         instant_retry = MHD_NO; /* client did not process everything */
       used -= processed;
       if (connection->have_chunked_upload == MHD_YES)
-        connection->current_chunk_offset += used;
+        connection->current_chunk_offset += (unsigned int)used;
       /* dh left "processed" bytes in buffer for next time... */
       buffer_head += used;
       available -= used;
